Fix uint8_t overflow in NUMERIC precision/scale arithmetic
resolve_numeric_typmod_from_exp extracted precision/scale with a 16-bit
  mask but stored them in uint8_t, so values above 255 wrapped and produced an
  incorrect result typmod. Widen the precision/scale locals to int32 and add
  max-precision arithmetic coverage to NUMERIC_TYPMOD.

+++ contrib/babelfishpg_tsql/src/pltsql_coerce.c
@@ -1502,13 +1502,13 @@ resolve_numeric_typmod_from_exp(Plan *plan, Node *expr, bool *found)
 						arg2type = InvalidOid;
 				int32		typmod1 = -1,
 						typmod2 = -1;
-				uint8_t		scale1,
+				int32		scale1,
 						scale2,
 						precision1,
 						precision2;
-				uint8_t		scale,
+				int32		scale,
 						precision;
-				uint8_t		integralDigitCount = 0;
+				int32		integralDigitCount = 0;
 				bool		found_typmod;
 				int32 		fixsize_default_typmod;
